'website_url'. Also, a signal is set up to automatically create a UserProfile when a new User is created.
In 'apps/users/admin.py', custom admin was established to extend the default User admin, adding a 'Task Count' field in the list display to show count of tasks assigned to each user.

These changes are made to provide a richer profile for users and to provide an enhanced admin view for user management.