"Add download functionality for shared files
Introduced a new function in the file_sharing views to handle file downloads. A download link was also added beside each file in the file_list template. Additionally, the URL for downloading a file was added to urls.py. Lastly, file_link attribute was added to the SharedFileAdmin class in the admin.py file allowing the file link to be displayed in the Django admin interface. All these changes foster a smoother user experience by facilitating quick file downloads."